+++ Application/Users/Queries/GetUserById.cs
@@ -5,13 +5,13 @@ using MediatR;
 
 namespace Application.Users.Queries;
 
-public record GetUserByIdQuery(int Id) : IRequest<UserDto?>;
+public record GetUserByIdQuery(uint Id) : IRequest<UserDto?>;
 
 public class GetUserByIdQueryValidator : AbstractValidator<GetUserByIdQuery>
 {
     public GetUserByIdQueryValidator()
     {
-        RuleFor(x => x.Id).GreaterThan(0).WithMessage("User ID must be greater than 0.");
+        RuleFor(x => (int)x.Id).GreaterThan(0).WithMessage("User ID must be greater than 0.");
     }
 }
 
@@ -19,6 +19,6 @@ public class GetUserByIdHandler(IUserReadRepository userRepository) : IRequestHa
 {
   public async Task<UserDto?> Handle(GetUserByIdQuery request, CancellationToken cancellationToken)
     {
-        return await userRepository.GetByIdAsync((uint)request.Id);
+        return await userRepository.GetByIdAsync(request.Id);
     }
 }
